Refactor: Type, ReservedManageCard 컴포넌트 분리 적용
